check if model exists and 404 before returning 403 errors

+++ src/Route/DeleteModelRoute.php
@@ -19,6 +19,14 @@ class DeleteModelRoute extends AbstractModelRoute
     {
         parent::buildResponse();
 
+        if (!$this->model->exists()) {
+            throw $this->modelNotFoundError();
+        }
+
+        if (!$this->hasPermission()) {
+            throw $this->permissionError();
+        }
+
         if ($this->model->delete()) {
             $this->response->setCode(204);
